[226] Moves tests to SortedMapTest from TreeMap
